refactor: extract Slack channel routes to routes/integrations/slack/channel.ts
