+++ assistant/src/runtime/routes/integration-routes.ts
@@ -1,11 +1,6 @@
 /**
  * Route handlers for integration config endpoints.
  *
- * Slack channel:
- * GET    /v1/integrations/slack/channel/config — get current config status
- * POST   /v1/integrations/slack/channel/config — validate and store credentials
- * DELETE /v1/integrations/slack/channel/config — clear credentials
- *
  * Channel verification (unified session API):
  * POST   /v1/channel-verification-sessions        — create session (inbound challenge, outbound verification, or trusted contact)
  * POST   /v1/channel-verification-sessions/resend  — resend outbound verification code
@@ -21,11 +16,6 @@ import {
   revokeVerificationForChannel,
   verifyTrustedContact,
 } from "../../daemon/handlers/config-channels.js";
-import {
-  clearSlackChannelConfig,
-  getSlackChannelConfig,
-  setSlackChannelConfig,
-} from "../../daemon/handlers/config-slack-channel.js";
 import { normalizePhoneNumber } from "../../util/phone.js";
 import { revokePendingSessions } from "../channel-verification-service.js";
 import { httpError } from "../http-errors.js";
@@ -37,40 +27,6 @@ import {
   startOutbound,
 } from "../verification-outbound-actions.js";
 import { verificationRateLimiter } from "../verification-rate-limiter.js";
-
-// ---------------------------------------------------------------------------
-// Slack channel config
-// ---------------------------------------------------------------------------
-
-/**
- * GET /v1/integrations/slack/channel/config
- */
-export function handleGetSlackChannelConfig(): Response {
-  const result = getSlackChannelConfig();
-  return Response.json(result);
-}
-
-/**
- * POST /v1/integrations/slack/channel/config
- *
- * Body: { botToken?: string, appToken?: string }
- */
-export async function handleSetSlackChannelConfig(
-  req: Request,
-): Promise<Response> {
-  const body = (await req.json()) as { botToken?: string; appToken?: string };
-  const result = await setSlackChannelConfig(body.botToken, body.appToken);
-  const status = result.success ? 200 : 400;
-  return Response.json(result, { status });
-}
-
-/**
- * DELETE /v1/integrations/slack/channel/config
- */
-export async function handleClearSlackChannelConfig(): Promise<Response> {
-  const result = await clearSlackChannelConfig();
-  return Response.json(result);
-}
 
 // ---------------------------------------------------------------------------
 // Channel verification (unified session API)
@@ -271,22 +227,6 @@ export async function handleRevokeVerificationBinding(
 
 export function integrationRouteDefinitions(): RouteDefinition[] {
   return [
-    // Slack
-    {
-      endpoint: "integrations/slack/channel/config",
-      method: "GET",
-      handler: () => handleGetSlackChannelConfig(),
-    },
-    {
-      endpoint: "integrations/slack/channel/config",
-      method: "POST",
-      handler: async ({ req }) => handleSetSlackChannelConfig(req),
-    },
-    {
-      endpoint: "integrations/slack/channel/config",
-      method: "DELETE",
-      handler: () => handleClearSlackChannelConfig(),
-    },
     // Channel verification (unified session API)
     {
       endpoint: "channel-verification-sessions",
